splits "test_journalist.py" into multiple files
